requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
tiktoken
faiss-cpu
python-multipart
jinja2
numpy
requests
langchain
langchain-community>=0.0.235
langchain-openai>=0.1.0
langchain-text-splitters
duckduckgo-search

langgraph>=0.0.40
langchain-core>=0.2.0

evidently>=0.4.9
whylabs-client>=0.8.5
psycopg2-binary
sqlalchemy>=2.0.0

mcp>=1.0.0

transformers>=4.30.0
torch>=2.0.0
datasets>=2.14.0
peft>=0.6.0
accelerate>=0.20.0
bitsandbytes>=0.41.0

streamlit>=1.28.0
plotly>=5.15.0

pydantic>=2.0.0
pyyaml>=6.0
python-json-logger>=2.0.0
prometheus-client>=0.17.0
